Renames the deprecated integration_installation webhook event.
