+++ api/package.json
@@ -7,7 +7,7 @@
     "dev": "node --inspect out/index.js",
     "compile": "tsc -watch -p ./",
     "build": "tsc",
-    "type-check": "tsc --noEmit --pretty",
+    "type-check": "tsc --noEmit --pretty --skipLibCheck",
     "postinstall": "yarn build; if [ $DATABASE_JSON_FILE ]; then yarn run setup; fi",
     "test": "jest && tslint source/**/**/**.ts",
     "lint": " tslint source/**/**/**.ts",
